Corregido error al varios jobs juntos. No se estaba teniendo en cuenta el resultado de ejecutar el job anterior. La expansión se realizaba antes de hora. Corregido error al interpretar variables: echo $A$B Ahora podemos ejecutar comandos, como:
export A=minishell.h; export B=minishell_b.h ; echo $A$B
